feat: create event
TODO fix wrong hour event and displayed our

+++ src/manager/callback.manager.ts
@@ -29,30 +29,43 @@ export class CallbackManager {
         const deleteTaskListRegex: RegExp = new RegExp('delete/[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}/task-list');
         const deleteEventRegex: RegExp = new RegExp('d/.*/e');
         const dateRegexp: RegExp = new RegExp('calendar-telegram-date-.*');
-        const alsoDateRegexp: RegExp = new RegExp('calendar-telegram-ignore-([12]\\\d{3}-(0[1-9]|1[0-2])-(0[1-9]|[12]\\\d|3[01]))');
+        const alsoDateRegexp: RegExp = new RegExp('calendar-telegram-ignore-([12]\\d{3}-(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01]))');
         const nextDateRegexp: RegExp = new RegExp('calendar-telegram-next-.*');
         const prevDateRegexp: RegExp = new RegExp('calendar-telegram-prev-.*');
         const hourRegexp: RegExp = new RegExp('^[0-2][0-9]:[0-5][0-9]$');
+        const durationRegexp: RegExp = new RegExp('duration/.*');
 
         // view task list
         if (viewRegex.test(context.event.payload)) {
+            if (!this.userStillExistant(context)) {
+                return Promise.resolve();
+            }
             await this.getTaskList(context.event.payload, context);
             return Promise.resolve();
         }
         // complete task
         if (completeTaskRegex.test(context.event.payload)) {
+            if (!this.userStillExistant(context)) {
+                return Promise.resolve();
+            }
             await this.completeTask(context.event.payload, context);
             return Promise.resolve();
         }
 
         // delete task
         if (deleteTaskListRegex.test(context.event.payload)) {
+            if (!this.userStillExistant(context)) {
+                return Promise.resolve();
+            }
             await this.deleteTaskList(context.event.payload, context);
             return Promise.resolve();
         }
 
         // delete event
         if (deleteEventRegex.test(context.event.payload)) {
+            if (!this.userStillExistant(context)) {
+                return Promise.resolve();
+            }
             const eventId: string = context.event.payload.split('/')[1];
             await ggl.deleteEvent(eventId, context.state.user.googleEmail, context.state.user.googleCredential.access_token);
             await context.sendMessage('Evento eliminado correctamente');
@@ -61,26 +74,21 @@ export class CallbackManager {
 
         // pickHour
         if (hourRegexp.test(context.event.payload)) {
-            // const hour: string = context.event.payload.split(':')[0];
-            // const min: string = context.event.payload.split(':')[1];
+            if (!this.userStillExistant(context)) {
+                return Promise.resolve();
+            }
+            const hour: string = context.event.payload.split(':')[0];
+            const min: string = context.event.payload.split(':')[1];
 
-            // const date = context.state.event.date;
-            // date.setHours(hour);
-            // date.setMinutes(min);
-
-            // await ggl.createEvent(state.user.googleCredential.access_token, state.user.googleEmail, state.event.summary, date);
-            // context.setState({
-            //     currentStatus: {
-            //         creatingEvent: false
-            //     }
-            // });
-            // await context.sendMessage('Creado');
+            const dateWithHourAndMinutes = context.state.event.date;
+            dateWithHourAndMinutes.setHours(hour);
+            dateWithHourAndMinutes.setMinutes(min);
 
             context.setState({
                 event: {
                     summary: context.state.event.summary,
                     location: context.state.event.location,
-                    date: context.state.event.date,
+                    date: dateWithHourAndMinutes,
                     duration: context.state.event.duration,
                     description: context.state.event.description,
                     hourAndMin: context.event.payload
@@ -92,16 +100,58 @@ export class CallbackManager {
                     inline_keyboard: durations
                 }
             });
-
         }
 
-        // pickDuration
-        if (1 === 1) {
+        // pick duration and create event
+        if (durationRegexp.test(context.event.payload)) {
+            if (!this.userStillExistant(context)) {
+                return Promise.resolve();
+            }
+            const durationMinutes: string = context.event.payload.split('/')[1];
+            const durationMinutesNumber: number = +durationMinutes;
+            const date = context.state.event.date;
+            const location = context.state.event.location;
+            const description = context.state.event.description;
+            const summary = context.state.event.summary;
 
+            if (_.isNil(durationMinutesNumber) || _.isNil(date) || _.isNil(location) || _.isNil(summary)) {
+                await context.sendMessage('Parece que algo ha ido mal . . . ¿Porque no lo intentas de nuevo?');
+                return Promise.resolve();
+            }
+            console.log('CURRENT EVENT:' + context.state.event);
+            try {
+                if (context.event.payload === 'duration/all_day') {
+                    await ggl.createEvent(
+                        context.state.user.googleCredential.access_token,
+                        context.state.user.googleEmail,
+                        summary,
+                        date,
+                        description,
+                        location
+                    );
+                    return Promise.resolve();
+                }
+                await ggl.createEvent(
+                    context.state.user.googleCredential.access_token,
+                    context.state.user.googleEmail,
+                    summary,
+                    date,
+                    description,
+                    location,
+                    durationMinutesNumber
+                );
+            } catch (error) {
+                await context.sendMessage(error.message);
+            }
+            await context.sendMessage('Evento creado correctamente');
+            return Promise.resolve();
         }
 
         // pickDate
         if (dateRegexp.test(context.event.payload) || alsoDateRegexp.test(context.event.payload)) {
+            if (!this.userStillExistant(context)) {
+                return Promise.resolve();
+            }
             const year: string = context.event.payload.split('-')[3];
             const month: string = context.event.payload.split('-')[4];
             const day: string = context.event.payload.split('-')[5];
@@ -111,6 +161,7 @@ export class CallbackManager {
                     location: context.state.event.location,
                     duration: context.state.event.duration,
                     summary: context.state.event.summary,
+                    description: context.state.event.description,
                     date: new Date(year + '-' + month + '-' + day)
                 }
             });
@@ -125,6 +176,9 @@ export class CallbackManager {
 
         // prevDate and next date
         if (prevDateRegexp.test(context.event.payload) || nextDateRegexp.test(context.event.payload)) {
+            if (!this.userStillExistant(context)) {
+                return Promise.resolve();
+            }
             const year: string = context.event.payload.split('-')[3];
             const month: string = context.event.payload.split('-')[4];
             const day: string = context.event.payload.split('-')[5];
@@ -137,7 +191,13 @@ export class CallbackManager {
             const maxDate = new Date();
             maxDate.setMonth(date.getMonth() + 12);
             // maxDate.setDate(date.getDate());
-            context.sendMessage('Selecciona la fecha del evento: ', calendar.setMinDate(date).setMaxDate(maxDate).getCalendar());
+            context.sendMessage(
+                'Selecciona la fecha del evento: ',
+                calendar
+                    .setMinDate(date)
+                    .setMaxDate(maxDate)
+                    .getCalendar()
+            );
             return Promise.resolve();
         }
 
@@ -194,25 +254,40 @@ export class CallbackManager {
                 break;
 
             case 'create_task_list':
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 state.currentStatus.creatingTaskList = true;
                 await context.sendMessage(Messages.TASK_LIST_NAME);
                 break;
 
             case 'dont_create_task_list':
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 await context.sendMessage(Messages.DONT_CREATE_TASKLIST);
                 break;
 
             case 'add_task':
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 state.currentStatus.addingTask = true;
                 await context.sendMessage(Messages.ADD_TASK);
                 break;
 
             case 'dont_add_task':
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 state.currentStatus.addingTask = false;
                 await context.sendMessage(Messages.TASK_LIST_DONE);
                 break;
 
             case 'task_list':
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 const taskLists: ITaskList[] = await Utils.getTasksLists(state.user);
                 if (_.isEmpty(taskLists)) {
                     await context.sendMessage(Messages.EMPTY_TASKS, {
@@ -256,6 +331,9 @@ export class CallbackManager {
                 break;
 
             case 'dropbox_settings':
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 const authUrl = dbx.getAuthUrl();
                 await context.sendMessage('Ajustes cuenta de Dropbox', {
                     reply_markup: {
@@ -278,6 +356,9 @@ export class CallbackManager {
                 break;
 
             case 'unlink_dropbox':
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 const auxUser: IUser = {
                     id: context.state.user.id,
                     username: context.state.user.username,
@@ -291,11 +372,17 @@ export class CallbackManager {
                 break;
 
             case 'change_username':
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 await context.sendMessage(Messages.CHANGE_USERNAME);
                 state.currentStatus.changingUsername = true;
                 break;
 
             case 'change_password':
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 await context.sendMessage(Messages.VALIDATE_CHANGE_PASSWORD);
                 state.currentStatus.validatingChangePassword = true;
                 break;
@@ -306,6 +393,9 @@ export class CallbackManager {
                 await context.sendMessage(Messages.LOGOUT);
                 break;
             case 'see_events':
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 const nextEvents: IGoogleEvent[] = await ggl.getEvents(state.user.googleCredential.access_token, state.user.googleEmail, 10);
                 if (_.isNil(nextEvents)) {
                     await context.sendMessage('Parece que no hay eventos disponibles');
@@ -327,7 +417,9 @@ export class CallbackManager {
                 }
                 break;
             case 'create_event':
-
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 context.sendMessage('¿Como se llamará el evento?');
                 context.setState({
                     currentStatus: {
@@ -339,7 +431,9 @@ export class CallbackManager {
                 break;
 
             case 'ignore_description':
-
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 context.setState({
                     event: {
                         summary: context.state.event.summary,
@@ -356,14 +450,23 @@ export class CallbackManager {
                 const maxDate = new Date();
                 maxDate.setMonth(today.getMonth() + 12);
                 maxDate.setDate(today.getDate());
-                context.sendMessage('Selecciona la fecha del evento: ', calendar.setMinDate(minDate).setMaxDate(maxDate).getCalendar());
+                context.sendMessage(
+                    'Selecciona la fecha del evento: ',
+                    calendar
+                        .setMinDate(minDate)
+                        .setMaxDate(maxDate)
+                        .getCalendar()
+                );
 
                 break;
 
             case 'add_description':
-
+                if (!this.userStillExistant(context)) {
+                    return Promise.resolve();
+                }
                 context.setState({
                     currentStatus: {
+                        creatingEvent: true,
                         insertingEventDescription: true
                     }
                 });
@@ -476,5 +579,13 @@ export class CallbackManager {
             .execute();
         await context.sendMessage(Messages.TASK_COMPLETED);
         return Promise.resolve();
+    }
+
+    private static async userStillExistant(context: any): Promise<boolean> {
+        if (_.isNil(context.state.user)) {
+            await context.sendMessage('Vaya, parece que me he olvidado de quien eres, ¿Porque no me refrescas la memoria?, utiliza /start para ello');
+            return Promise.resolve(false);
+        }
+        return Promise.resolve(true);
     }
 }
